LPS-56211 move and rename
